+++ src/services/fetchers/index.ts
@@ -1,0 +1,6 @@
+import CryptoComparePriceFetcher from './CryptoComparePriceFetcher';
+import CryptoCompareHistoHourFetcher from './CryptoCompareHistoHourFetcher';
+import CryptoCompareHistoDayFetcher from './CryptoCompareHistoDayFetcher';
+import GVolImpliedVolatilityFetcher from './GVolImpliedVolatilityFetcher';
+
+export {CryptoComparePriceFetcher, CryptoCompareHistoDayFetcher, CryptoCompareHistoHourFetcher, GVolImpliedVolatilityFetcher};